move dfsa collection for queries to new QueryLinter
