+++ tests/unit/source_code/test_jobs.py
@@ -16,7 +16,6 @@ from databricks.labs.ucx.source_code.python_libraries import PythonLibraryResolv
 from databricks.labs.ucx.source_code.known import KnownList
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.service import compute, jobs, pipelines
-from databricks.sdk.service.sql import Query, ListQueryObjectsResponseQuery
 from databricks.sdk.service.workspace import ExportFormat
 
 from databricks.labs.ucx.source_code.linters.files import FileLoader, ImportFileResolver
@@ -539,39 +538,3 @@ def test_full_lineage_is_converted_to_json():
         '{"object_type": "path", "object_id": "abc"}, '
         '{"object_type": "path", "object_id": "xyz"}]'
     )
-
-
-@pytest.mark.parametrize(
-    "name, query, dfsa_paths, is_read, is_write",
-    [
-        ("simple", "SELECT * from dual", [], False, False),
-        (
-            "location",
-            "CREATE TABLE hive_metastore.indices_historical_data.sp_500 LOCATION 's3a://db-gtm-industry-solutions/data/fsi/capm/sp_500/'",
-            ["s3a://db-gtm-industry-solutions/data/fsi/capm/sp_500/"],
-            False,
-            True,
-        ),
-    ],
-)
-def test_workflow_linter_collects_dfsas_from_queries(
-    name,
-    query,
-    dfsa_paths,
-    is_read,
-    is_write,
-    mock_path_lookup,
-    simple_dependency_resolver,
-    empty_index,
-):
-    ws = create_autospec(WorkspaceClient)
-    crawlers = create_autospec(DirectFsAccessCrawlers)
-    query = Query.from_dict({"parent_path": "workspace", "display_name": name, "query_text": query})
-    response = ListQueryObjectsResponseQuery.from_dict(query.as_dict())
-    ws.queries.list.return_value = iter([response])
-    linter = WorkflowLinter(ws, simple_dependency_resolver, mock_path_lookup, empty_index, crawlers)
-    dfsas = linter.collect_dfsas_from_queries()
-    crawlers.assert_not_called()
-    assert set(dfsa.path for dfsa in dfsas) == set(dfsa_paths)
-    assert not any(dfsa for dfsa in dfsas if dfsa.is_read != is_read)
-    assert not any(dfsa for dfsa in dfsas if dfsa.is_write != is_write)
